Fixed a bug with auto-starting the video when the view size != the video size.
TextureVideoView currently assumes that the surface size will be identical to the video size.  This is a naive assumption: the video size is only a default (a hint) to the texture view; the layout may size the texture view however it likes resulting in a different surface size (eg. match_parent).